+++ CHANGELOG.md
@@ -17,7 +17,9 @@ bottom of the file for a change log entry template.
   commands to the documentation file
 * C implementation files
 * A directory for a CDH board "monolithic" TAB example
-* Adds utilities and scripts supporting C examples
+* Utilities and scripts supporting C examples
+* A Python implementation file
+* Directories for Python RX and TX examples
 
 ### Fixed
 * Fixes Markdown table styling in the documentation file